requirements.txt
fastapi==0.97.0
uvicorn==0.27.0.post1
pydantic==1.10.15
pandas
numpy
oracledb
scikit-learn
matplotlib
seaborn
river
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
